requirements.txt
lxml==6.0.2
MouseInfo==0.1.3
pillow==12.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-docx==1.2.0
pytweening==1.2.0
typing_extensions==4.15.0
